+++ delphi/scorers/__init__.py
@@ -2,7 +2,7 @@ from .classifier.detection import DetectionScorer
 from .classifier.fuzz import FuzzingScorer
 from .embedding.embedding import EmbeddingScorer
 from .scorer import Scorer
-from .simulator.oai_simulator import OpenAISimulator
+from .simulator.simulator import OpenAISimulator
 from .surprisal.surprisal import SurprisalScorer
 
 __all__ = [
