--- delphi/scorers/simulator/oai_autointerp/__init__.py
@@ -1,13 +1,0 @@
-from .activations import ActivationRecord
-from .explanations import (
-    ExplanationNeuronSimulator,
-    LogprobFreeExplanationTokenSimulator,
-    simulate_and_score,
-)
-
-__all__ = [
-    "ActivationRecord",
-    "ExplanationNeuronSimulator",
-    "LogprobFreeExplanationTokenSimulator",
-    "simulate_and_score",
-]
