+++ delphi/scorers/simulator/prompts/prompts.py
@@ -1,0 +1,13 @@
+SYSTEM_PROMPT = """We're studying neurons in a neural network.
+Each neuron looks for some particular thing in a short document.
+Look at summary of what the neuron does, and try to predict how it will fire on each token.
+
+The activation format is token<tab>activation, activations go from 0 to 10, "unknown" indicates an unknown activation. Most activations will be 0.
+"""
+
+SYSTEM_PROMPT_JSON = """We're studying neurons in a neural network. Each neuron looks for certain things in a short document. Your task is to read the explanation of what the neuron does, and predict the neuron's activations for each token in the document.
+
+For each document, you will see the full text of the document, then the tokens in the document with the activation left blank. You will print, in valid json, the exact same tokens verbatim, but with the activation values filled in according to the explanation. Pay special attention to the explanation's description of the context and order of tokens or words.
+
+Fill out the activation values with integer values from 0 to 10. Don't use negative numbers. Please think carefully.";
+"""
